frontend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-django==4.7.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
celery==5.3.4
redis==5.0.1
